+++ src/Pages/Admin/categoryManagement/Category.jsx
@@ -9,6 +9,7 @@ import { logoutAdmin } from "@/redux/adminSlice";
 import { useDispatch } from "react-redux";
 import ReactPaginate from "react-paginate";
 import { Calendar, Search } from "lucide-react";
+import AlertDialogueButton from "@/Components/AlertDialogueButton";
 const Category = () => {
   const [name, setName] = useState("");
   const [description, setDescription] = useState("");
@@ -172,46 +173,46 @@ const Category = () => {
   };
 
   const handleRestore = async (id) => {
-    const result = confirm("Are you sure to restore categorie");
+    // const result = confirm("Are you sure to restore categorie");
     try {
-      if (result) {
-        const response = await axios.patch(`/admin/categoryRestore/${id}`);
-        Toast.fire({
-          icon: "success",
-          title: `${response.data.message}`,
-        });
-        setCategories((prev) =>
-          prev?.map((cat) =>
-            cat?._id === id ? { ...cat, isDeleted: false } : cat
-          )
-        );
+      // if (result) {
+      const response = await axios.patch(`/admin/categoryRestore/${id}`);
+      Toast.fire({
+        icon: "success",
+        title: `${response.data.message}`,
+      });
+      setCategories((prev) =>
+        prev?.map((cat) =>
+          cat?._id === id ? { ...cat, isDeleted: false } : cat
+        )
+      );
 
-        // window.location.reload();
-      }
+      // window.location.reload();
+      // }
     } catch (error) {
       alert(error.response?.data.message);
     }
   };
   const handleDelete = async (id) => {
-    const result = confirm("Are you sure to delete categorie");
+    // const result = confirm("Are you sure to delete categorie");
     try {
-      if (result) {
-        const response = await axios.delete(`/admin/category/${id}`);
-        setCategories((prev) =>
-          prev?.map((cat) =>
-            cat?._id === id ? { ...cat, isDeleted: true } : cat
-          )
-        );
-        Toast.fire({
-          icon: "success",
-          title: `${response.data.message}`,
-        });
-      } else {
-        Toast.fire({
-          icon: "success",
-          title: `Cancelled the deletion`,
-        });
-      }
+      // if (result) {
+      const response = await axios.delete(`/admin/category/${id}`);
+      setCategories((prev) =>
+        prev?.map((cat) =>
+          cat?._id === id ? { ...cat, isDeleted: true } : cat
+        )
+      );
+      Toast.fire({
+        icon: "success",
+        title: `${response.data.message}`,
+      });
+      // } else {
+      //   Toast.fire({
+      //     icon: "success",
+      //     title: `Cancelled the deletion`,
+      //   });
+      // }
     } catch (error) {
       Toast.fire({
         icon: "success",
@@ -420,24 +421,30 @@ const Category = () => {
                     )}
 
                     {category.isDeleted ? (
-                      <button
+                      <AlertDialogueButton
+                        name={
+                          <button className="rounded bg-green-600 p-1 font-mono">
+                            Restore
+                          </button>
+                        }
                         onClick={() => handleRestore(category._id)}
-                        className="rounded bg-green-600 p-1 font-mono"
-                      >
-                        Restore
-                      </button>
+                        className="text-white px-4 py-2 bg-gray-700 hover:bg-gray-600 rounded shadow"
+                      />
                     ) : (
-                      <i
-                        className="fas fa-trash-alt"
+                      <AlertDialogueButton
+                        name={<i className="fas fa-trash-alt"></i>}
                         onClick={() => handleDelete(category._id)}
-                      ></i>
+                        className="text-white px-4 py-2 bg-gray-700 hover:bg-gray-600 rounded shadow"
+                      />
                     )}
                   </td>
                 </tr>
               ))
             ) : (
               <tr className="border-t">
-                <td className="p-2 font-bold mt-10 ms-10 ">Loading ....</td>
+                <td className="p-2 font-bold mt-10 ms-10 ">
+                  No Categories added yet!
+                </td>
               </tr>
             )}
           </tbody>
